Refactor `ParamConfigManager` to `backtesting` Package and Update Dependencies (#827)
